requirements.txt
beautifulsoup4==4.3.2
Django==1.7.7
django-appconf==1.0.1
django-compressor==1.4
django-debug-toolbar==1.3.0
django-filter==0.9.2
django-filters==0.1.0
django-redis==4.0.0
django-stdimage==1.2.2
django-zurb-foundation==5.5.0
djangorestframework==3.1.2
geopy==1.10.0
ipython==3.0.0
mysqlclient==1.3.6
nose==1.3.0
numpy==1.7.1
oauthlib==0.7.2
Pillow==2.7.0
progressbar2==2.7.3
redis==2.10.3
requests==2.6.0
requests-oauthlib==0.4.2
six==1.3.0
sqlparse==0.1.15
twython==3.2.0
